pyproject.toml
[tool.poetry]
name = "toxic_bot"
version = "0.1.0"
description = "Testing how to build discord bots"
authors = ["Tapan Pandita <tapan.pandita@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"
python-dotenv = "^0.15.0"
aiohttp = "^3.6.3"
aiodns = "^2.0.0"
cchardet = "^2.1.7"
ipython = "^7.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
